+++ src/Ext/TAction/Body.h
@@ -18,6 +18,7 @@ enum class PhobosTriggerAction : unsigned int
 	BinaryOperation = 504,
 	RunSuperWeaponAtLocation = 505,
 	RunSuperWeaponAtWaypoint = 506,
+	MessageForSpecifiedHouse = 9931,
 };
 
 class TActionExt
@@ -59,6 +60,7 @@ public:
 	ACTION_FUNC(BinaryOperation);
 	ACTION_FUNC(RunSuperWeaponAtLocation);
 	ACTION_FUNC(RunSuperWeaponAtWaypoint);
+	ACTION_FUNC(MessageForSpecifiedHouse);
 
 	static bool RunSuperWeaponAt(TActionClass* pThis, int X, int Y);